+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+## [1.0.15](https://github.com/tamnguyenvan/screenarc/compare/v1.0.14...v1.0.15) (2025-10-14)
+
+
+### Bug Fixes
+
+* export video issue ([4cc5bfa](https://github.com/tamnguyenvan/screenarc/commit/4cc5bfac2f18c14d57d133473e2b4deb12cc77e4))
+
+## [1.0.14](https://github.com/tamnguyenvan/screenarc/compare/v1.0.13...v1.0.14) (2025-10-14)
+
+
+### Bug Fixes
+
+* app icon ([e81844f](https://github.com/tamnguyenvan/screenarc/commit/e81844fe1173208058ac5aa34d430936ded4da62))
+* debian icon ([ea6f2ef](https://github.com/tamnguyenvan/screenarc/commit/ea6f2ef68ad1e6abccccee13c1dab5fbfe43033d))
+
+## [1.0.13](https://github.com/tamnguyenvan/screenarc/compare/v1.0.12...v1.0.13) (2025-10-14)
+
+
+### Bug Fixes
+
+* build .deb, .rpm & .zip for Linux ([4673f6b](https://github.com/tamnguyenvan/screenarc/commit/4673f6bffc858f62be3e4572c8cd57fb6dabcca6))
+
+## [1.0.12](https://github.com/tamnguyenvan/screenarc/compare/v1.0.11...v1.0.12) (2025-10-14)
+
+
+### Bug Fixes
+
+* build .deb, .rpm & .zip for Linux ([7d38058](https://github.com/tamnguyenvan/screenarc/commit/7d38058677bc5ff48e6c5d45158a6273fd32892d))
+
 # Changelog
 
 All notable changes to this project will be documented in this file.